Updated bug fixes.
